# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus Hosting API
  version: 1.0.0
  description: |
    Production-grade federated static site hosting network.
    Every node exposes this API. Authentication uses session cookies (browser)
    or Bearer API tokens (CLI / CI). Inter-node requests use Ed25519 signatures.
  contact:
    name: The No Hands Company
    url: https://github.com/The-No-Hands-company/Nexus-Hosting
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: Local node API
...
paths:
...
  /sites/{id}/domains:
    get:
      operationId: listCustomDomains
      tags: [domains]
      summary: List custom domains for a site
      security:
        - sessionAuth: []
        - bearerToken: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Custom domains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomDomain"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: addCustomDomain
      tags: [domains]
      summary: Add a custom domain and get DNS verification instructions
      security:
        - sessionAuth: []
        - bearerToken: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain]
              properties:
                domain:
                  type: string
                  example: "www.mysite.com"
      responses:
        "201":
          description: Domain added with DNS instructions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomDomainWithInstructions"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /domains/{id}/verify:
    post:
      operationId: verifyCustomDomain
      tags: [domains]
      summary: Trigger a DNS TXT verification check for a domain
      security:
        - sessionAuth: []
        - bearerToken: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                  domain:
                    type: string
                  status:
                    type: string
                  lastError:
                    type: ["string", "null"]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /domains/{id}:
    delete:
      operationId: deleteCustomDomain
      tags: [domains]
      summary: Remove a custom domain from a site
      security:
        - sessionAuth: []
        - bearerToken: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Domain removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── Analytics ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: sid
      description: Browser session cookie set by /login flow
    bearerToken:
      type: http
      scheme: bearer
      description: API token created via POST /tokens (prefix fh_)
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
        code:
          type: string
        status:
          type: integer
...
    CustomDomain:
      type: object
      required: [id, siteId, domain, verificationToken, status, createdAt]
      properties:
        id:
          type: integer
        siteId:
          type: integer
        domain:
          type: string
        verificationToken:
          type: string
        status:
          type: string
          enum: [pending, verified, failed]
        verifiedAt:
          type: ["string", "null"]
        lastCheckedAt:
          type: ["string", "null"]
        lastError:
          type: ["string", "null"]
        createdAt:
          type: string
    CustomDomainWithInstructions:
      allOf:
        - $ref: "#/components/schemas/CustomDomain"
        - type: object
          properties:
            instructions:
              type: object
              properties:
                txt:
                  type: object
                  properties:
                    name:
                      type: string
                    type:
                      type: string
                    value:
                      type: string
                cname:
                  type: object
                  properties:
                    name:
                      type: string
                    type:
                      type: string
                    value:
                      type: string
...